Move files
Update example in its own folder
