+++ README.md
@@ -3,4 +3,4 @@
 This is a simple task manager.  It is meant to be used with a barebones project.
 
 Build example with
-gcc -0 main.c task_mgr.c
+gcc -o example main.c ../src/task_mgr.c
